requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
google-adk>=1.20.0
google-generativeai>=0.8.0
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
numpy>=1.24.0
av>=10.0.0
soundfile>=0.12.0
websockets>=11.0